update ts-utils (use int type)

+++ packages/utils/ts-utils/src/array/access/first.ts
@@ -1,8 +1,9 @@
+import { uint32 } from '../../types';
 import { ReadonlyNonEmptyArray } from '../non-empty-array';
 import { at } from './at';
 
 export function first<T>(array: ReadonlyNonEmptyArray<T>): T;
 export function first<T>(array: readonly T[]): T | undefined;
 export function first<T>(array: readonly T[]): T | undefined {
-  return at(array, 0);
+  return at(array, 0 as uint32);
 }
