+++ packages/utils/ts-utils/src/array/slice/take-last.test.ts
@@ -1,32 +1,33 @@
+import { uint32 } from '../../types';
 import { testArrayEquality } from '../array-tester';
 import { takeLast } from './take-last';
 
 testArrayEquality({
   testName: 'takeLast',
-  target: takeLast([1, 2, 3, 4, 5], 3),
+  target: takeLast([1, 2, 3, 4, 5], 3 as uint32),
   toBe: [3, 4, 5],
 });
 
 testArrayEquality({
   testName: 'takeLast',
-  target: takeLast([1, 2, 3, 4, 5], 0),
+  target: takeLast([1, 2, 3, 4, 5], 0 as uint32),
   toBe: [],
 });
 
 testArrayEquality({
   testName: 'takeLast',
-  target: takeLast([1, 2, 3, 4, 5], -1),
+  target: takeLast([1, 2, 3, 4, 5], -1 as uint32),
   toBe: [],
 });
 
 testArrayEquality({
   testName: 'takeLast',
-  target: takeLast([1, 2, 3, 4, 5], 5),
+  target: takeLast([1, 2, 3, 4, 5], 5 as uint32),
   toBe: [1, 2, 3, 4, 5],
 });
 
 testArrayEquality({
   testName: 'takeLast',
-  target: takeLast([1, 2, 3, 4, 5], 6),
+  target: takeLast([1, 2, 3, 4, 5], 6 as uint32),
   toBe: [1, 2, 3, 4, 5],
 });
