+++ packages/utils/ts-utils/src/array/slice/take.test.ts
@@ -1,28 +1,29 @@
+import { uint32 } from '../../types';
 import { testArrayEquality } from '../array-tester';
 import { take } from './take';
 
 testArrayEquality({
   testName: 'take',
-  target: take([1, 2, 3, 4, 5], 3),
+  target: take([1, 2, 3, 4, 5], 3 as uint32),
   toBe: [1, 2, 3],
 });
 testArrayEquality({
   testName: 'take',
-  target: take([1, 2, 3, 4, 5], 0),
+  target: take([1, 2, 3, 4, 5], 0 as uint32),
   toBe: [],
 });
 testArrayEquality({
   testName: 'take',
-  target: take([1, 2, 3, 4, 5], -1),
+  target: take([1, 2, 3, 4, 5], -1 as uint32),
   toBe: [],
 });
 testArrayEquality({
   testName: 'take',
-  target: take([1, 2, 3, 4, 5], 5),
+  target: take([1, 2, 3, 4, 5], 5 as uint32),
   toBe: [1, 2, 3, 4, 5],
 });
 testArrayEquality({
   testName: 'take',
-  target: take([1, 2, 3, 4, 5], 6),
+  target: take([1, 2, 3, 4, 5], 6 as uint32),
   toBe: [1, 2, 3, 4, 5],
 });
